moved ReadSecrets to bulker, tests
