+++ internal/pkg/bulk/secret.go
@@ -2,7 +2,7 @@
 // or more contributor license agreements. Licensed under the Elastic License;
 // you may not use this file except in compliance with the Elastic License.
 
-package dl
+package bulk
 
 import (
 	"encoding/json"
